MedScan_Doctor_Prescription-reader/requirements.txt
pytesseract>=0.3.8
pillow>=9.0.0
transformers>=4.20.0
torch>=1.10.0
spacy>=3.4.0
fastapi>=0.75.0
uvicorn>=0.17.0
streamlit>=1.10.0
python-multipart>=0.0.5
requests>=2.28.0
smtplib>=0.0.1
python-dotenv>=0.20.0